### 2025-03-31 #7959 @editorialbot
[PRE REVIEW]: OpenHPS: A Modular Framework to Facilitate the Development of FAIR Positioning Systems
**Submitting author:** @maximvdw (<a href="http://orcid.org/0000-0003-0314-7107">Maxim Van de Wynckel</a>)
**Repository:** https://github.com/OpenHPS/openhps-core
**Branch with paper.md** (empty if default branch): maxim
**Version:** v1.0.4
