+++ docs/paper/paper.md
@@ -14,7 +14,7 @@ authors:
     orcid: 0000-0001-9916-0837
     affiliation: 1
 affiliations:
- - name: Vrije Universiteit Brussel, Web & Information Systems Engineer Lab, Brussels, Belgium
+ - name: Vrije Universiteit Brussel, Web & Information Systems Engineering Lab, Brussels, Belgium
    index: 1
 date: 31 March 2025
 bibliography: paper.bib
@@ -36,7 +36,7 @@ All concepts, ranging from positions to sensor values, can be expressed in vario
 
 ![OpenHPS graph of a positioning system](images/openhps-example-2.png){width=60%}
 
-Any data pushed through the graph can be serialised to JSON or RDF data using the POSO ontology [@vandewynckel2022iswc], enabling interoperability between systems. Each node in the graph represent a step in the processing of data frames from source to sink as shown in Figure 1. A `SourceNode` generates information, a `ProcessingNode` processes information and a `SinkNode` consumes information.
+Any data pushed through the graph can be serialised to JSON or RDF data using the POSO ontology [@vandewynckel2022iswc], enabling interoperability between systems. Each node in the graph represents a step in the processing of data frames from source to sink as shown in Figure 1. A `SourceNode` generates information, a `ProcessingNode` processes information and a `SinkNode` consumes information.
 
 OpenHPS is modular by design, mainly due to the ability to extend data frames and objects. These extensions allow for different data objects, including different sensors like cameras, IMU sensors or spatial landmarks such as Bluetooth beacons. Each node in a graph can be extended as well, allowing the creation of custom algorithms that can be added or removed from a positioning system. Researchers can focus on the prototyping of new algorithms without having to worry about the integration of these algorithms into a larger system.
 
@@ -44,11 +44,11 @@ OpenHPS is modular by design, mainly due to the ability to extend data frames an
 Our framework is designed to offer FAIR positioning systems, leveraging the open approach of the framework and the data these systems produce. With our modules such as `@openhps/rdf` and `@openhps/solid`, we have enabled developers to serialise their data to RDF and store it in Solid Pods. This facilitates the creation of positioning systems that are transparent and privacy preserving, while also ensuring that other positioning systems or consumer applications can access the data, regardless of whether they are created using OpenHPS. Extensions such as `@sembeacon/openhps` provide support for semantic beacons [@vandewynckel2023], enabling the discovery (i.e. findability) of these systems and the data they produce. 
 
 ### Performance
-Since our framework uses TypeScript, it uses JavaScript at runtime, which is single-threaded by default. To overcome the challenges associated with creating real-time stream processing systems, our graphs or portions of the graph can be executed using web workers. All data transmitted through our graph is serialisable, eliminating the need for developers to handle this serialisation or communication between web workers themselves.
+Since our framework uses TypeScript, it uses JavaScript at runtime, which is single-threaded by default. To overcome the challenges associated with creating real-time stream processing systems, our graphs or portions of the graph can be executed using Web Workers. All data transmitted through our graph is serialisable, eliminating the need for developers to handle this serialisation or communication between Web Workers themselves.
 
 In addition to the ability to run graphs on multiple workers, our communication nodes such as MQTT, combined with the serialisability of data, allows developers to offload the processing of complex tasks to other servers or dedicated processors. For more high-demanding algorithms as used in computer vision and visual SLAM, modules such as `@openhps/opencv` and `@openhps/openvslam` create C++ bindings in NodeJS.
 
 # Examples of Research Work
[…]
