### 2025-04-02 #7959 @Maximvdw
Suggested reviewers: [Abdelhak Bougouffa](https://reviewers.joss.theoj.org/reviewers/2305) (positioning and sensor fusion, but no TS or JS as languages), [Allie Tatarian](https://reviewers.joss.theoj.org/reviewers/3539) (interoperability and linked data, JS -> mainly for the FAIR aspect of the software), [jorgedch](https://reviewers.joss.theoj.org/reviewers/2112) (positioning, including some out-of-the-box positioning which fits the hybrid part of openhps, JS), [Sunny Patel](https://reviewers.joss.theoj.org/reviewers/11395) (distributed systems, TS, stream processing)
