### 2025-04-04 #7959 @danielskatz
@sunnypatel165 - You can see JOSS published papers on https://joss.theoj.org From there, you can click on any of them, and see near the top right a link to the paper review.

See https://joss.readthedocs.io/en/latest/reviewer_guidelines.html for how a JOSS review works, and https://joss.readthedocs.io/en/latest/review_criteria.html for the items you would be working to review.

JOSS seeks to complete reviews in 2-4 weeks, but can take longer.
