### 2025-04-18 #7959 @danielskatz
👋 @gareth-cross & @swilcock0 & @gautam-sharma1, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
