### 2025-04-26 #7959 @floe
As discussed with @danielskatz on Mastodon, I'd be available for a review. Do I understand correctly that I need to open a new issue for the review itself?
