### 2025-05-05 #8113 @floe
Sorry for the noise, was looking for the reminder feature... are there more extensive docs for editorialbot somewhere?
