### 2025-05-05 OpenHPS/openhps-core#296 @floe
Suggestion: additional related work for paper
I'm currently reviewing this paper for JOSS (https://github.com/openjournals/joss-reviews/issues/8113), and have an extra suggestion for related work: "Automatic configuration of pervasive sensor networks for augmented reality" by Pustka et al., https://ieeexplore.ieee.org/document/5482592 (disclaimer: I'm one of the authors, and the similarity between OpenHPS and UbiTrack was what prompted me to review in the first place). In particular, UbiTrack also uses a dataflow graph to process and transform tracking data between coordinate systems, so there is indeed some overlap present.
