### 2025-05-05 OpenHPS/openhps-core#297 @floe
Suggestion: discuss extensibility in paper
I'm reviewing this paper for JOSS (https://github.com/openjournals/joss-reviews/issues/8113), and I have a question about the system design: perhaps I have some outdated conceptions here, but I would generally design the core of the system as a native library (in C, C++, or Rust) and then add bindings to other languages - for example, I would love to test-drive OpenHPS with Python, but as far as I can see, that wouldn't be currently possible? Could you add 1-2 sentences to the paper to explain how your design decisions went in this context?
