### 2025-05-05 #8113 @danielskatz
> Sorry for the noise, was looking for the reminder feature... are there more extensive docs for editorialbot somewhere?

Does https://joss.readthedocs.io/en/latest/editorial_bot.html help?
