### 2025-05-05 OpenHPS/openhps-core#298 @floe
Suggestion: discuss relation to ROS?
[cf. JOSS review in https://github.com/openjournals/joss-reviews/issues/8113 ]

On reading a bit more on the overall context, I was starting to wonder what the main differences between OpenHPS and ROS are? Based on my experience, most people pick ROS for various tracking tasks, so why should they rather go with OpenHPS instead? This also relates to how well Typescript can handle the relatively hard realtime requirements that many tracking tasks have.
