### 2025-05-07 OpenHPS/openhps-core#298 @Maximvdw
Hey! First of all, thank you for taking your time to review the framework.

The design philosophy of OpenHPS is to be a framework that supports a wide variety of platforms, use cases and technologies. To achieve this, we formalised what a _hybrid_ positioning looks like, both with the data structure and pipeline of information that gets processed. This formalised data structure enables us to optionally decentralise nodes over multiple platforms, threads or other systems. Essentially, we do not want to limit ourselves to indoor positioning, robotics, navigation, ... but also other use cases, such as, for example, tracking objects on a tabletop.

ROS is similar to that in a way, it also has _nodes_ and is decentralised. A ROS package is a set of nodes linked together. Each node handles part of a task (in a package) and can be chained to other packages. In a sense, you can compare a [graph](https://openhps.org/docs/core/interfaces/graph) shape with ROS packages, each package consisting of internal nodes. One year after our release of OpenHPS [`rosjslib`](https://github.com/RobotWebTools/roslibjs) was released which also enables the creation of nodes on the Web of hybrid applications.
![ROS example nodes](https://wiki.ros.org/indoor_localization?action=AttachFile&do=get&target=package_setup.png)

The main difference and issue that we found between ROS and OpenHPS is the flexibility of reusing these nodes for different use cases and their platform-independent implementation, enabling implementation of parts of the processing on a mobile phone (hybrid application or web app), server (nodejs) or other services. Packages provide a set of nodes, but we found that nodes within those packages are often hard to reuse due to very specific or different data formats (requiring a node to transform data) and the method how nodes are connected requires custom remapping.

With OpenHPS, both the communication in-between nodes is flexible (MQTT, REST, sockets), the serialization is flexible ([RDF](https://github.com/OpenHPS/openhps-rdf), [protobuf](https://github.com/OpenHPS/openhps-protobuf)), and each node is reusable for any use case that you want. On the one hand, it requires more work to create a positioning system as opposed to using the 'indoor localization' package of ROS with a built-in graph. But on the other hand, the flexibility of having each node available and compatible with each other makes it very flexible.

As for the performance. For low-level data and processing, we rely on C-bindings to low-level frameworks such as OpenCV (https://github.com/OpenHPS/openhps-opencv). Instead of moving this raw data from node to node, we would in that case move pointers around.

Finally, last but definitely not least, related to the data formalisation - we also provide developers with the opportunity to use standardised exporting of linked data (RDF-XML, Turtle, JSON-LD) with modules such as https://github.com/OpenHPS/openhps-rdf , which makes data created by OpenHPS not only modular within the framework, but also outside.

Feel free to ask any follow-up questions!
