### 2025-05-08 OpenHPS/openhps-core#296 @Maximvdw
Hey, I am looking into it and will come back tomorrow with maybe some additional questions.
