### 2025-05-08 OpenHPS/openhps-core#297 @Maximvdw
Hey, I briefly touched on the subject in your other ticket (https://github.com/OpenHPS/openhps-core/issues/298). With OpenHPS, we wanted to focus on the reusability of data and the reusability of algorithms. Our primary focus was to create a modular framework that is not specific to a single use case. After investigating many different use cases (AR, indoor positioning, robotics, context-aware applications, crowd management, security, ....) and technologies, we found that the best solution to cover these use cases would be to develop a framework that could run on the Web, mobile applications and servers. For our data reusability, we then relied on open standards such as GeoSPARQL and the [Semantic Sensor Network Ontology](https://www.w3.org/TR/vocab-ssn/).

I agree, from a technical & performance point of view, creating the framework in a low-level language facilitates the integration in other languages - but this requires an additional step for creating bindings.

Testing OpenHPS as a framework in Python is indeed not possible. That being said, since our communication protocols are modular as well ([MQTT](https://github.com/OpenHPS/openhps-mqtt), REST, ...) the pushing and retrieval of data can be done in any other language.

---

I will modify the paper to make this design philosophy and methodology more clear. At the moment, we kept the statement of need short due to the limited length of 1000 words.
