### 2025-05-08 fda33e9
chore: paper update based on initial feedback
