+++ docs/paper/paper.bib
@@ -87,3 +87,35 @@ doi={10.1007/978-3-031-60635-9_14}
     url={https://github.com/schollz/find3},
     year={2020}
 }
+
+@ARTICLE{5482592,
+  author={Pustka, Daniel and Huber, Manuel and Waechter, Christian and Echtler, Florian and Keitler, Peter and Newman, Joseph and Schmalstieg, Dieter and Klinker, Gudrun},
+  journal={IEEE Pervasive Computing}, 
+  title={{Automatic Configuration of Pervasive Sensor Networks for Augmented Reality}}, 
+  year={2011},
+  volume={10},
+  number={3},
+  pages={68-79},
+  doi={10.1109/MPRV.2010.50}
+}
+
+@INPROCEEDINGS{7125601,
+  author={García, Enrique and Poudereux, Pablo and Hernández, Álvaro and Ureña, Jesús and Gualda, David},
+  booktitle={2015 IEEE International Conference on Industrial Technology (ICIT)}, 
+  title={{A Robust UWB Indoor Positioning System for Highly Complex Environments}}, 
+  year={2015},
+  volume={},
+  number={},
+  pages={3386-3391},
+  doi={10.1109/ICIT.2015.7125601}
+}
+
+@inproceedings{quigley2009ros,
+	title={{ROS: An Open-Source Robot Operating System}},
+	author={Quigley, Morgan and Conley, Ken and Gerkey, Brian and Faust, Josh and Foote, Tully and Leibs, Jeremy and Wheeler, Rob and Ng, Andrew Y},
+	booktitle={Proceedings of the International Workshop on Open Source Software},
+	address={Kobe, Japan},
+	year={2009},
+	month=may,
+  url={https://api.semanticscholar.org/CorpusID:6324125}
+}
