### 2025-05-08 #8113 @Maximvdw
_Checking word count_

Based on the first feedback, the new article proof applies the following changes (https://github.com/OpenHPS/openhps-core/commit/fda33e98ad1587d724494a289c03fe2d1e000279#diff-a0198f95473849eaebe4f3b238ede8ea1049d03f0ec252df8cf7f25845eca217)

The latest version can be found in this post: https://github.com/openjournals/joss-reviews/issues/8113#issuecomment-2863581261
