### 2025-05-08 OpenHPS/openhps-core#296 @Maximvdw
From a processing and flexibility aspect, I indeed think that Ubitrack overlaps with OpenHPS. I can see from this paper why you also wanted to learn more about the real-time processing capabilities of the framework. For a use case such as AR, using TypeScript is indeed less ideal. However, if you consider modern use cases with for example, WebXR, OpenHPS becomes useful to enable absolute positioning in such a context. There are also examples in OpenHPS where we use AR (https://openhps.org/posts/2024/03/26/).

In OpenHPS, we focus less on the context in environments and more on the context of the processing of data - so that the pipeine/dataflow is indeed also described. From what I can gather, this pipeline is described using the proprietary UTQL configuration that is then distributed to each local dataflow? This is quite interesting, does that mean that the graph/pipeline **has** to be constructed on the Ubitrack server, or can extensions be added without being preconfigured on the main server?

> As some background that does not 100% relate to the paper, we use the POSO ontology (https://poso.openhps.org) within our @openhps/rdf component to provide context about the procedures.

> Creating spatial relationship graphs is possible in OpenHPS through [@openhps/geospatial](https://github.com/OpenHPS/openhps-geospatial/tree/master/src/data) and the mapping to linked data in [@openhps/rdf](https://github.com/OpenHPS/openhps-rdf/tree/master/src/mapping/geospatial). However, it requires more effort from developers as it is not of the primary context providers.
