### 2025-05-09 OpenHPS/openhps-core#296 @floe
> In OpenHPS, we focus less on the context in environments and more on the context of the processing of data - so that the pipeine/dataflow is indeed also described. From what I can gather, this pipeline is described using the proprietary UTQL configuration that is then distributed to each local dataflow? This is quite interesting, does that mean that the graph/pipeline **has** to be constructed on the Ubitrack server, or can extensions be added without being preconfigured on the main server?

You're right, Ubitrack is quite centralized, and that did make some sense for the industrial scenarios we were mostly targeting at that time - but in retrospect, this approach (plus the proprietary UTQL description) were not the best design decisions, mostly because it made the whole thing very inflexible.
