### 2025-05-12 OpenHPS/openhps-core#296 @floe
Oh, just realized this can be closed now. Thanks!
