### 2025-05-13 #8113 @floe
Mostly done with the paper and general parts, still need to go through the software parts of the checklist.
