### 2025-05-27 OpenHPS/openhps-core#301 @parakhsid
Suggestion: Performance Benchmarking
The Performance section (line 63-67), claims improvements from web workers but lacks evidence. Adding a bench mark on sinlge v/s multi-threaded execution would help strengthen the claim.

(Line68-72), There’s mention of using external processors and C++ bindings, however consider a comparison table showing CPU-bound NodeJS implementation vs native module performance.
