### 2025-05-28 OpenHPS/openhps-core#301 @Maximvdw
Hi, first of all, thank you for reviewing the paper and framework. We have a benchmark https://github.com/OpenHPS/openhps-core/tree/master/test/bench but the results are indeed not added to the short paper. I will see if I can add the original benchmark from a report we published a few years ago in the paper itself (while still adhering to the word count). As an FYI, the benchmark is on page 11 of https://openhps.org/publications/2020/techreport/ or worst case reference it from the paper itself.

> (Line68-72), There’s mention of using external processors and C++ bindings, however consider a comparison table showing CPU-bound NodeJS implementation vs native module performance.
The external processors are only for C++ libraries (e.g. OpenCV in https://github.com/OpenHPS/openhps-opencv). It will not be possible to comapre this to nodejs performance since this does not exist. However, I can clarify in the paper that the bindings are for third-party libraries that process data
