### 2025-05-28 #8113 @Maximvdw
How strict is the 1000 wordcount? Currently I have 1042 from what I can see from the editorbot, should I try to lower this?
