+++ docs/paper/paper.bib
@@ -119,3 +119,12 @@ doi={10.1007/978-3-031-60635-9_14}
 	month=may,
   url={https://api.semanticscholar.org/CorpusID:6324125}
 }
+
+@techreport{mvdewync2020openhps,
+	title={{OpenHPS: An Open Source Hybrid Positioning System}},
+	author={{Van de Wynckel}, Maxim and Signer, Beat},
+  institution={Vrije Universiteit Brussel},
+	year={2020},
+	number={\mbox{WISE-2020-01}},
+	doi={10.48550/ARXIV.2101.05198}
+}
