+++ docs/paper/paper.md
@@ -16,39 +16,39 @@ authors:
 affiliations:
  - name: Vrije Universiteit Brussel, Web & Information Systems Engineering Lab, Brussels, Belgium
    index: 1
-date: 08 May 2025
+date: 29 May 2025
 bibliography: paper.bib
 ---
 
 ## Abstract
-Positioning systems determine the location of people and objects using various technologies and algorithms. While GPS dominates outdoor positioning, indoor and smaller-scale systems often require alternative technologies for improved latency, accuracy or efficiency. These systems are frequently developed as single-use prototypes with no standard data format, hindering reusability and expansion. OpenHPS addresses these challenges by providing a modular, graph-based framework for creating versatile positioning systems on multiple platforms. It supports a wide range of algorithms and allows developers to extend functionality by sharing custom nodes for sensor fusion and novel algorithms.
+Positioning systems determine the location of people and objects using various technologies and algorithms. While GPS dominates outdoor positioning, indoor and smaller-scale systems often require alternative technologies for improved latency, accuracy, or efficiency. These systems are often developed as single-use prototypes with no standard data format, hindering reusability and expansion. OpenHPS addresses these challenges by providing a modular, graph-based framework for creating versatile positioning systems on multiple platforms. It supports a wide range of algorithms and enables extensibility through custom nodes for sensor fusion and algorithm integration.
 
 ## Keywords
 Hybrid positioning systems, interoperable positioning systems, indoor positioning, stream processing, linked data, RDF
 
 # Statement of Need
-Existing positioning systems or frameworks [@georgiou2015anyplace][@find3][@5482592] often focus on specific use cases, platforms or a set of algorithms. While this facilitates the deployment of a system, it also limits the freedom to design and reuse custom algorithms or positioning systems. In systems developed for academic purposes [@7125601], proprietary systems are created that are difficult to replicate, expand or implement in a production-ready environment due to the use of software such as MatLab. Furthermore, the data produced by such systems is often unstandardised. OpenHPS was created to address these issues by providing a modular framework that allows developers and researchers to create positioning systems for a wide range of use cases. The framework and data created with our framework are designed to be interoperable, allowing developers to share their algorithms and data with others. Using additional modules that add support for Solid [@vandewynckel2022] and DHTs, we have enabled the development of findable, accessible, interoperable and reusable (FAIR) positioning systems.
+Existing positioning systems or frameworks [@georgiou2015anyplace;@find3;@5482592] often focus on specific use cases, platforms, or a set of algorithms. While this facilitates the deployment of a system, it limits the flexibility to design or reuse custom algorithms and systems. In systems developed for academic purposes [@7125601], proprietary systems are often difficult to replicate, extend, or deploy in a production-ready environment due to the use of software such as MatLab. Furthermore, the data produced by such systems is often unstandardised. OpenHPS was created to address these issues by providing a modular framework that allows developers and researchers to create positioning systems for a wide range of use cases. The framework and data created with our framework is designed to be interoperable, allowing developers to share their algorithms and data with others. Using additional modules that add support for Solid [@vandewynckel2022] and DHTs, we have enabled the development of findable, accessible, interoperable and reusable (FAIR) positioning systems.
 
 ## Framework Overview
[…]
