+++ docs/paper/paper.md
@@ -49,6 +49,6 @@ Our framework is designed to offer FAIR positioning systems, leveraging its open
 OpenHPS uses JavaScript at runtime. To overcome the challenges associated with creating real-time dataflows, our graph can be executed using Web Workers [@mvdewync2020openhps]. All data transmitted through our graph is serialisable, eliminating the need for developers to handle this serialisation or communication between workers. In addition to the ability to run graphs on multiple workers, communication nodes, such as MQTT, enable developers to offload the processing of complex tasks to other servers or dedicated processors. For more high-demanding algorithms, modules like `@openhps/opencv` and `@openhps/openvslam` provide C++ bindings to low-level libraries. While these bindings may not achieve the same performance, the ability to reuse other libraries is part of the main requirements.
 
 # Examples of Research Work
-OpenHPS has been a building block for various research that has been used in indoor positioning systems [@vandewynckel2021], and its ability to serialise location data to RDF has been demonstrated in an application that aims to preserve privacy and transparency using Solid Pods [@vandewynckel2022]. Further, OpenHPS was used in the SemBeacon demonstrator application [@vandewynckel2023] that is written in CapacitorJS and uses the framework to deserialise positioning data and positioning systems. Modules have been created for other academic projects, such as the FidMark ontology [@vandewynckel2024], which provides fiducial marker classification within the framework. OpenHPS's modular node design and domain focus facilitate the sharing of algorithms and findings, as well as the rapid creation of prototypes and demonstrators that make use of location data. Finally, OpenHPS and the related modules have been used to collect several datasets that can be used by researchers to evaluate their algorithms [@openhps2021dataset], [@maxim_van_de_wynckel_2025], [@maxim_van_de_wynckel_beat_signer_2025]. 
+OpenHPS has been a building block for various research that has been used in indoor positioning systems [@vandewynckel2021], and its ability to serialise location data to RDF has been demonstrated in an application that aims to preserve privacy and transparency using Solid Pods [@vandewynckel2022]. Further, OpenHPS was used in the SemBeacon demonstrator application [@vandewynckel2023] that is written in CapacitorJS and uses the framework to deserialise positioning data and positioning systems. Modules have been created for other academic projects, such as the FidMark ontology [@vandewynckel2024], which provides fiducial marker classification within the framework. OpenHPS's modular node design and domain focus facilitate the sharing of algorithms and findings, as well as the rapid creation of prototypes and demonstrators that make use of location data. Finally, OpenHPS and the related modules have been used to collect several datasets that can be used by researchers to evaluate their algorithms [@openhps2021dataset;@maxim_van_de_wynckel_2025;@maxim_van_de_wynckel_beat_signer_2025]. 
 
 # References
