### 2025-05-29 #8113 @Maximvdw
**Changelog**
- Based on https://github.com/OpenHPS/openhps-core/issues/301:
  - Added a reference to our 2020 technical report which contains a benchmark for web workers
  - Rephrased the "C++ bindings to frameworks" with "C++ bindings to **libraries**"
  - Clarified that C++ bindings promotes reusability and may not improve performance
- Added a figure to illustrate the data frames and data objects
- Fixed multi-cite
- Proofreading and minor changes
