### 2025-05-30 #8113 @danielskatz
> How strict is the 1000 wordcount? Currently I have 1042 from what I can see from the editorbot, should I try to lower this?

It is not strict. 1042 (or 1027) are both fine.
