+++ docs/paper/paper.md
@@ -27,7 +27,7 @@ Positioning systems determine the location of people and objects using various t
 Hybrid positioning systems, interoperable positioning systems, indoor positioning, stream processing, linked data, RDF
 
 # Statement of Need
-Existing positioning systems or frameworks [@georgiou2015anyplace;@find3;@5482592] often focus on specific use cases, platforms, or a set of algorithms. While this facilitates the deployment of a system, it limits the flexibility to design or reuse custom algorithms and systems. In systems developed for academic purposes [@7125601], proprietary systems are often difficult to replicate, extend, or deploy in a production-ready environment due to the use of software such as MatLab. Furthermore, the data produced by such systems is often unstandardised. OpenHPS was created to address these issues by providing a modular framework that allows developers and researchers to create positioning systems for a wide range of use cases. The framework and its produced data are designed to be interoperable, allowing developers to share their algorithms and data with others. Using additional modules that add support for Solid [@vandewynckel2022] and DHTs, we have enabled the development of findable, accessible, interoperable, and reusable (FAIR) positioning systems.
+Existing positioning systems or frameworks [@georgiou2015anyplace;@find3;@5482592] often focus on specific use cases, platforms, or a set of algorithms. While this facilitates the deployment of a system, it limits the flexibility to design or reuse custom algorithms and systems. In systems developed for academic purposes [@7125601], proprietary systems are often difficult to replicate, extend, or deploy in a production-ready environment due to the use of software such as MatLab. Furthermore, the data produced by such systems is often unstandardised. OpenHPS was created to address these issues by providing a modular data-centric framework that allows developers and researchers to create positioning systems for a wide range of use cases. The framework and its produced data are designed to be interoperable, enabling developers to share their algorithms and data with others. Using additional modules that add support for Solid [@vandewynckel2022] and DHTs, we have enabled the development of findable, accessible, interoperable, and reusable (FAIR) positioning systems.
 
 ## Framework Overview
 `OpenHPS` is an open-source hybrid positioning system framework written in TypeScript. It runs on the server (Node.js), in the browser, or within hybrid mobile applications. The general design of a positioning system created using OpenHPS consists of a graph with a set of nodes that process data. Background services can be added to the graph to persist data or to communicate with other systems. Our decision to use TypeScript compared to using a low-level language was based on the goal of achieving cross-platform reusability.
