+++ test/bench/worker.bench.ts
@@ -17,176 +17,198 @@ import { ComputingNode } from '../mock/nodes/ComputingNode';
 import * as Benchmark from 'benchmark';
 import * as fs from 'fs';
 
-const size = parseInt(process.argv[2]);
-const minSamples = parseInt(process.argv[3]);
-const workerCount = parseInt(process.argv[5]);
-const concurrency = parseInt(process.argv[4]);
+const sizes: number[] = [5000, 10000, 15000];
 
-console.log("Size: " + size);
-console.log("Min Samples: " + minSamples);
-console.log("Worker Count: " + workerCount);
-console.log("Concurrency: " + concurrency);
+let promise: Promise<void> = Promise.resolve();
 
-const frames: DataFrame[] = new Array();
-const suite = new Suite();
-const settings: Benchmark.Options = {
-    defer: true,
-    minSamples,
-    initCount: 10,
-    delay: 10
-};
-const settingsCreate: Benchmark.Options = {
-    defer: true,
-    minSamples: 1,
-    delay: 10
-};
-let model: Model | undefined;
+for (let i = 0 ; i < sizes.length ; i++) {
+    const size = sizes[i];
+    const minSamples = parseInt(process.argv[2]);
+    const workerCount = parseInt(process.argv[3]);
+    const concurrency = parseInt(process.argv[4]);
 
-async function init() {
-    for (let i = 0 ; i < 100 ; i++) {
-        const dummyFrame = new DataFrame();
-        const dummyObject = new DataObject("dummy", "Dummy Data Object");
-        const position = new Absolute3DPosition(0, 0, 0);
-        position.velocity.linear = new LinearVelocity(0.1, 0.1, 0.1);
-        position.velocity.angular = new AngularVelocity(0.1, 0.1, 0.1);
-        position.orientation = new Orientation(0, 0, 0, 1);
-        dummyObject.setPosition(position);
-        dummyFrame.source = dummyObject;
-        dummyFrame.addObject(dummyObject);
-        frames.push(dummyFrame);
-    }
-}
-
-function createModel(workers: number): Promise<Model> {
-    return new Promise((resolve, reject) => {
-        let model: Model;
-        ModelBuilder.create()
-            .addShape(GraphBuilder.create()
-                .from()
-                .via(new WorkerNode((builder, modelBuilder, args) => {
-                    const { ComputingNode } = require(path.join(__dirname, '../mock/nodes/ComputingNode'));
-                    builder.via(new ComputingNode(args.size));
-                }, {
-                    poolSize: workers,
-                    poolConcurrency: concurrency,
-                    directory: __dirname,
-                    services: [],
-                    args: {
-                        size
-                    }
-                }))
-                .to(new CallbackSinkNode(() => {
-                    
-                }, {
-                    uid: "sink"
-                })))
-            .build().then((m: Model) => {
-                model = m;
-                return model.push(new DataFrame());
-            }).then(() => {
-                resolve(model);
-            }).catch(ex => {
-                reject(ex);
-            });
+    promise = promise.then(() => {
+        return execute(size, minSamples, workerCount, concurrency);
     });
 }
 
-function testFunction(model: Model | undefined, deferred: any): void {
-    if (model === undefined) {
-        console.error("Model is undefined!");
-        return deferred.resolve();
-    }
-    let promises = new Array();
-    frames.forEach(frame => {
-        promises.push(model.push(frame));
-    });
-    const sink = model.findNodeByUID("sink") as CallbackSinkNode<any>;
-    let count = 0;
-    sink.callback = (frame) => {
-        count++;
-        if (count === frames.length) {
-            deferred.resolve();
-        }
-    };
-    Promise.resolve(promises);
-}
[…]
