### 2025-06-04 OpenHPS/openhps-core#302 @floe
Suggestion: add example to the front page
After going through the final JOSS checklist, I have one more request: I would like to see a simple producer-filter-consumer example, ready to run, on the front page. The current example is a bit too abstract IMHO to help a newcomer with using the framework. Something like the one in https://github.com/OpenHPS/openhps-opencv/blob/master/examples/browser-script/index.html would be nice.

(cf. https://github.com/openjournals/joss-reviews/issues/8113)
