### 2025-06-04 OpenHPS/openhps-core#302 @floe
To clarify: this can be more or less a test case, with artificial producers and consumers (or perhaps it would be possible to have a "mouse position" producer in the browser, then a Kalman filter, and a 2D plot as consumer? That would be ideal IMHO.)
