### 2025-06-04 OpenHPS/openhps-core#302 @Maximvdw
Thanks for the suggestion. I will create the mouse example with a custom active source node (mouse onmove implementation), sink node (chart.js output) and simple moving average as the filter for the X,Y position.

Will comment when it is published.
