### 2025-06-04 ec0e7c2
fix: sma node should not divide by taps when amount of items is less than the amount of taps
