### 2025-06-04 b565318
fix: sma node should not divide by taps when amount of items is less than the amount of taps
