### 2025-06-04 OpenHPS/openhps-core#302 @Maximvdw
Again, thank you for the mouse suggestion. The example is now available on the website:
https://openhps.org/

The example showcases
- Reference spaces (origin of onmouseevent is top left, for the chart it's bottom right)
- Active source node
- Sink node
- SMA node
while still keeping it minimal enough for the front page.

There are two buttons underneath the example, one to test the example and one that leads to a new page in the documentation with the source code of the custom nodes.

https://openhps.org/docs/tutorials/mouse/

![](https://cdn.maximvdw.be/chrome_OsTvh2J4sd.png)
![](https://cdn.maximvdw.be/chrome_GgI73lSMKY.gif)
