### 2025-06-04 c831646
docs: example from openhps site
