### 2025-06-04 OpenHPS/openhps-core#302 @Maximvdw
I've linked the example in the usage section
