### 2025-06-04 68281c2
chore: benchmark results and script
