### 2025-06-05 f88c46f
Update paper.bib
