+++ docs/paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{vandewynckel2021,
-  title={{Indoor Positioning Using the OpenHPS Framework}},
+  title={Indoor Positioning Using the {OpenHPS} Framework},
   author={{Van de Wynckel}, Maxim and Signer, Beat},
   booktitle={Proceedings of the 11th International Conference on Indoor Positioning and Indoor Navigation (IPIN 2021)},
   pages={1--8},
@@ -8,7 +8,7 @@
 }
 
 @inproceedings{vandewynckel2023,
-  title={{SemBeacon: A Semantic Proximity Beacon Solution for Discovering and Detecting the Position of Physical Things}},
+  title={{SemBeacon}: A Semantic Proximity Beacon Solution for Discovering and Detecting the Position of Physical Things},
   author={Van de Wynckel, Maxim and Signer, Beat},
   booktitle={Proceedings of the 13th International Conference on the Internet of Things (IoT 2023)},
   year=2023,
@@ -16,7 +16,7 @@
 }
 
 @inproceedings{vandewynckel2022,
-  title={{A Solid-based Architecture for Decentralised Interoperable Location Data}},
+  title={A Solid-based Architecture for Decentralised Interoperable Location Data},
   author={{Van de Wynckel}, Maxim and Signer, Beat},
   booktitle={Proceedings of the 12th International Conference on Indoor Positioning and Indoor Navigation (IPIN 2022), CEUR Workshop Proceedings},
   year={2022}, 
@@ -26,7 +26,7 @@
 }
 
 @inproceedings{georgiou2015anyplace,
-  title={{Anyplace: A Crowdsourced Indoor Information Service}},
+  title={Anyplace: A Crowdsourced Indoor Information Service},
   author={Georgiou, Kyriakos and Constambeys, Timotheos and Laoudias, Christos and Petrou, Lambros and Chatzimilioudis, Georgios and Zeinalipour-Yazti, Demetrios},
   booktitle={Proceedings of the International Conference on Mobile Data Management: Systems, Services and Middleware (MDM 2015)},
   pages={291--294},
@@ -38,7 +38,7 @@
 author="Van de Wynckel, Maxim
 and Valadez, Isaac
 and Signer, Beat",
-title={{FidMark: A Fiducial Marker Ontology for Semantically Describing Visual Markers}},
+title={{FidMark}: A Fiducial Marker Ontology for Semantically Describing Visual Markers},
 booktitle="Proceedings of The Semantic Web (ESWC 2024)",
 year="2024",
 publisher = {Springer-Verlag},
@@ -50,7 +50,7 @@ doi={10.1007/978-3-031-60635-9_14}
 @InProceedings{vandewynckel2022iswc,
     author="Van de Wynckel, Maxim
     and Signer, Beat",
-    title={{POSO: A Generic Positioning System Ontology}},
+    title={POSO: A Generic Positioning System Ontology},
     booktitle={Proceedings of the 21st International Semantic Web Conference (ISWC 2022)},
     year={2022},
     address={Virtual conference},
@@ -58,7 +58,7 @@ doi={10.1007/978-3-031-60635-9_14}
 }
 
 @misc{maxim_van_de_wynckel_2025,
-	title={{Garage Positioning Dataset}},
+	title={Garage Positioning Dataset},
 	DOI={10.34740/KAGGLE/DS/6654647},
 	publisher={Kaggle},
 	author={Maxim {Van de Wynckel}},
@@ -70,19 +70,19 @@ doi={10.1007/978-3-031-60635-9_14}
 	doi = {10.34740/KAGGLE/DS/6760212},
 	year = {2025},
[…]
