### 2025-06-05 #8113 @danielskatz
@Maximvdw - Your paper text looks good. I note there is no acknowledgments section. Should there be?

Also, the references are formatted (manually) incorrectly for JOSS.  Please merge https://github.com/OpenHPS/openhps-core/pull/303 or let me know what you disagree with.
