### 2025-06-05 #8113 @Maximvdw
As the project/research is not funded, no acknowledgements are needed. I've checked the PR (thanks for making the required changes), apart from one project name it is fine.
