+++ docs/paper/paper.bib
@@ -16,7 +16,7 @@
 }
 
 @inproceedings{vandewynckel2022,
-  title={A Solid-based Architecture for Decentralised Interoperable Location Data},
+  title={A {S}olid-based Architecture for Decentralised Interoperable Location Data},
   author={{Van de Wynckel}, Maxim and Signer, Beat},
   booktitle={Proceedings of the 12th International Conference on Indoor Positioning and Indoor Navigation (IPIN 2022), CEUR Workshop Proceedings},
   year={2022}, 
