### 2025-06-05 #8113 @Maximvdw
> 👋 [@Maximvdw](https://github.com/Maximvdw) - At this point could you:
>
> * Make a tagged release of your software, and list the version tag of the archived version here.
> * Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.

- v1.0.13  https://github.com/OpenHPS/openhps-core/releases/tag/v1.0.13

- [10.5281/zenodo.15602601](https://doi.org/10.5281/zenodo.15602601)
