### 2025-06-05 #8113 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
