### 2025-06-05 #8113 @Maximvdw
> [@editorialbot](https://github.com/editorialbot) recommend-accept
>
> [@Maximvdw](https://github.com/Maximvdw) - please take a look at the output document from this, and ensure that you don't see any problems.

I've checked the final article (PDF) and it looks ok
